linux-pipewire: Factor out formats list into a file
This will help share these formats with the upcoming Camera portal
code.
